Move ProgressBar to flat structure
